Improve Spring Session sample application
This commit improves Spring Session sample application by providing multiple build profiles that make it possible to easily try out different session stores. By default, JDBC session store backed by an in-memory embedded H2 database is used.
