+++ spring-boot-samples/pom.xml
@@ -71,7 +71,7 @@
 		<module>spring-boot-sample-secure-oauth2-actuator</module>
 		<module>spring-boot-sample-secure-oauth2-resource</module>
 		<module>spring-boot-sample-servlet</module>
-		<module>spring-boot-sample-session-redis</module>
+		<module>spring-boot-sample-session</module>
 		<module>spring-boot-sample-simple</module>
 		<module>spring-boot-sample-test</module>
 		<module>spring-boot-sample-test-nomockito</module>
