+++ spring-boot-samples/spring-boot-sample-session/src/main/resources/application.properties
